refactor(client): extract useTransferAnalytics from P2PTransfer
